% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePlotLiP.R
\name{BarcodePlotLiP}
\alias{BarcodePlotLiP}
\title{Barcode plot. Shows protein coverge of LiP modified peptides.}
\usage{
BarcodePlotLiP(
  data,
  fasta,
  which.prot = "all",
  width = 12,
  height = 4,
  address = ""
)
}
\arguments{
\item{data}{list of data.tables containing LiP and TrP data in MSstatsLiP
format. Can be the output of converters such as
\code{\link[MSstatsLiP]{SpectronauttoMSstatsLiPFormat}}.}

\item{fasta}{A string of path to a FASTA file}

\item{which.prot}{a list of peptides to be visualized.}

\item{width}{width of the saved file. Default is 10.}

\item{height}{height of the saved file. Default is 10.}

\item{address}{the name of folder that will store the results. Default
folder is the current working directory. The other assigned folder has to
be existed under the current working directory. An output pdf file is
automatically created with the default name of "VolcanoPlot.pdf" or
"Heatmap.pdf". The command address can help to specify where to store the
file as well as how to modify the beginning of the file name. If
address=FALSE, plot will be not saved as pdf file but showed in window}
}
\value{
plot or pdf
}
\description{
Barcode plot. Shows protein coverge of LiP modified peptides.
}
\examples{
Add example
}
