% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlotLiP.R
\name{calculate.pc}
\alias{calculate.pc}
\title{Internal function to run Principle Component Analysis on a dataset. Returns
results of PCA.}
\usage{
calculate.pc(data, center.pca, scale.pca)
}
\description{
Internal function to run Principle Component Analysis on a dataset. Returns
results of PCA.
}
