% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummarizationLiP.R
\name{dataSummarizationLiP}
\alias{dataSummarizationLiP}
\title{Summarizes LiP and TrP datasets seperately using methods from MSstats.
Returns list of two summarized datasets}
\usage{
dataSummarizationLiP(
  data,
  logTrans = 2,
  normalization = "equalizeMedians",
  normalization.LiP = FALSE,
  nameStandards = NULL,
  nameStandards.LiP = NULL,
  address = "",
  fillIncompleteRows = TRUE,
  featureSubset = "all",
  featureSubset.LiP = "all",
  remove_uninformative_feature_outlier = FALSE,
  remove_uninformative_feature_outlier.LiP = FALSE,
  n_top_feature = 3,
  n_top_feature.LiP = 3,
  summaryMethod = "TMP",
  equalFeatureVar = TRUE,
  censoredInt = "NA",
  cutoffCensored = "minFeature",
  MBimpute = TRUE,
  MBimpute.LiP = FALSE,
  remove50missing = FALSE,
  fix_missing = NULL,
  maxQuantileforCensored = 0.999,
  clusters = NULL
)
}
\description{
Summarizes LiP and TrP datasets seperately using methods from MSstats.
Returns list of two summarized datasets
}
