% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{LiPRawData}
\alias{LiPRawData}
\title{LiPRawData}
\format{
A data.table consisting of 546 rows and 29 columns. Raw LiP data for use in testing
and examples.
}
\usage{
LiPRawData
}
\description{
Example of input LiP dataset.
}
\details{
Input to MSstatsLiP converter SpectronauttoMSstatsLiPFormat.
Contains the following columns:

\itemize{
\item R.Condition : Label of conditions (EG Disease/Control)
\item R.FileName : Name of spectral processing run
\item R.Replicate : Name of biological replicate
\item PG.ProteinAccessions : Protein name
\item PG.ProteinGroups : Protein name, can be multiple
\item PG.Quantity : Protein Quantity
\item PEP.GroupingKey : Peptide grouping
\item PEP.StrippedSequence : Peptide sequence
\item PEP.Quantity : Peptide quantity
\item EG.iRTPredicted : Predicted value
\item EG.Library : Name of library
\item EG.ModifiedSequence : Peptide sequence including any post-translational modifications
\item EG.PrecursorId : Peptide sequence wiht modifications including charge
\item EG.Qvalue : Qvalue
\item FG.Charge : Identified Ion charge
\item FG.Id : Peptide sequence with charge
\item FG.PrecMz : Prec Mz reading
\item FG.Quantity : Initial quantity reading
\item F.Charge : F.Charge
\item F.FrgIon : Fragment ion
\item F.FrgLossType : Label for loss type
\item F.FrgMz : Mz reading
\item F.FrgNum : numeric Frg
\item F.FrgType : character label for Frg
\item F.ExcludedFromQuantification : True/False boolean for if to exclude
\item F.NormalizedPeakArea : Normalized peak intensity
\item F.NormalizedPeakHeight : Normalized peak height
\item F.PeakArea : Unnormalized peak area
\item F.PeakHeight : Unnormalized peak height
}
}
\examples{
head(LiPRawData)

}
\keyword{datasets}
