% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinetoMSstatsLiPFormat.R
\name{SkylinetoMSstatsLiPFormat}
\alias{SkylinetoMSstatsLiPFormat}
\title{Converts raw LiP MS data from Skyline into the format needed for
MSstatsLiP.}
\usage{
SkylinetoMSstatsLiPFormat(LiP.data, TrP.data = NULL)
}
\arguments{
\item{LiP.data}{name of LiP Spectronaut output, which is long-format.}

\item{TrP.data}{name of TrP Spectronaut output, which is long-format.}

\item{use_log_file}{logical. If TRUE, information about data processing
will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be
added to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing will be
printed to the console.}

\item{log_file_path}{character. Path to a file to which information about
data processing will be saved.
If not provided, such a file will be created automatically.
If \code{append = TRUE}, has to be a valid path to a file.}

\item{base}{start of the file name.}
}
\description{
Takes as as input both raw LiP and Trp outputs from Skyline.
}
\examples{

#fasta_path <- "../inst/extdata/ExampleFastaFile.fasta"

#MSstatsLiP_data <- SkylinetoMSstatsLiPFormat(LiPRawData,
#                                              fasta_path,
#                                              TrPRawData)

}
