% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePTM.R
\name{tidyFasta}
\alias{tidyFasta}
\title{Read and tidy a FASTA file}
\usage{
tidyFasta(path)
}
\arguments{
\item{path}{a string of path pointing towards a fasta file}
}
\description{
reads and tidys FASTA file.
}
\examples{
tidyFasta("https://www.uniprot.org/uniprot/P02407.fasta")
}
