% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePlotLiP.R
\name{BarcodePlotLiP}
\alias{BarcodePlotLiP}
\title{Barcode plot. Shows protein coverge of LiP modified peptides.}
\usage{
BarcodePlotLiP(data, fasta, which.prot = "all")
}
\description{
Barcode plot. Shows protein coverge of LiP modified peptides.
}
