% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonLiP.R
\name{groupComparisonLiP}
\alias{groupComparisonLiP}
\title{Model LiP and TrP data and make adjustments if needed
Returns list of three modeled datasets}
\usage{
groupComparisonLiP(
  data,
  contrast.matrix = "pairwise",
  moderated = FALSE,
  adj.method = "BH"
)
}
\description{
Model LiP and TrP data and make adjustments if needed
Returns list of three modeled datasets
}
