% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonPlotsLiP.R
\name{groupComparisonPlotsLiP}
\alias{groupComparisonPlotsLiP}
\title{Visualization for model-based analysis and summarization}
\usage{
groupComparisonPlotsLiP(
  data = data,
  type = type,
  sig = 0.05,
  FCcutoff = FALSE,
  logBase.pvalue = 10,
  ylimUp = FALSE,
  ylimDown = FALSE,
  xlimUp = FALSE,
  x.axis.size = 10,
  y.axis.size = 10,
  dot.size = 3,
  text.size = 4,
  text.angle = 0,
  legend.size = 13,
  ProteinName = TRUE,
  colorkey = TRUE,
  numProtein = 100,
  clustering = "both",
  width = 10,
  height = 10,
  which.Comparison = "all",
  which.Protein = "all",
  which.Models = "all",
  address = ""
)
}
\description{
Visualization for model-based analysis and summarization
}
