% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlotLiP.R
\name{PCAPlotLiP}
\alias{PCAPlotLiP}
\title{Plot PCA for LiP and TrP datasets. Takes as input LiP data and outputs plot.}
\usage{
PCAPlotLiP(
  data,
  center.pca = TRUE,
  scale.pca = TRUE,
  n.components = 10,
  bar.plot = TRUE,
  protein.pca = TRUE,
  comparison.pca = FALSE,
  which.pep = "all",
  which.comparison = "all",
  width = 10,
  height = 10,
  address = ""
)
}
\description{
Plot PCA for LiP and TrP datasets. Takes as input LiP data and outputs plot.
}
