% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylinetoMSstatsLiPFormat.R
\name{SkylinetoMSstatsLiPFormat}
\alias{SkylinetoMSstatsLiPFormat}
\title{Converts raw LiP MS data from Skyline into the format needed for
MSstatsLiP.}
\usage{
SkylinetoMSstatsLiPFormat(LiP.data, TrP.data = NULL)
}
\arguments{
\item{LiP.data}{name of LiP Spectronaut output, which is long-format.}

\item{TrP.data}{name of TrP Spectronaut output, which is long-format.}
}
\value{
a \code{list} of two data.frames in \code{MSstatsLiP} format
}
\description{
Takes as as input both raw LiP and Trp outputs from Skyline.
}
\examples{

## Output will be in format
head(MSstatsLiP_data[["LiP"]])
head(MSstatsLiP_data[["TrP"]])
}
