% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectronauttoMSstatsLiPFormat.R
\name{SpectronauttoMSstatsLiPFormat}
\alias{SpectronauttoMSstatsLiPFormat}
\title{Converts raw LiP MS data from Spectronautt into the format needed for
MSstatsLiP. Needs as input both LiP Trp outputs from Spectronautt.}
\usage{
SpectronauttoMSstatsLiPFormat(
  LiP.data,
  Trp.data,
  fasta,
  annotation = NULL,
  intensity = "PeakArea",
  filter_with_Qvalue = TRUE,
  qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  which.Conditions = "all"
)
}
\description{
Converts raw LiP MS data from Spectronautt into the format needed for
MSstatsLiP. Needs as input both LiP Trp outputs from Spectronautt.
}
