% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectronauttoMSstatsLiPFormat.R
\name{SpectronauttoMSstatsLiPFormat}
\alias{SpectronauttoMSstatsLiPFormat}
\title{Converts raw LiP MS data from Spectronautt into the format needed for
MSstatsLiP.}
\usage{
SpectronauttoMSstatsLiPFormat(
  LiP.data,
  fasta,
  Trp.data = NULL,
  annotation = NULL,
  intensity = "PeakArea",
  filter_with_Qvalue = TRUE,
  qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  fewMeasurements = "remove",
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  which.Conditions = "all"
)
}
\arguments{
\item{LiP.data}{name of LiP Spectronaut output, which is long-format.}

\item{fasta}{A string of path to a FASTA file, used to match LiP peptides.}

\item{Trp.data}{name of TrP Spectronaut output, which is long-format.}

\item{annotation}{name of 'annotation.txt' data which includes Condition,
BioReplicate, Run. If annotation is already complete in Spectronaut, use
annotation=NULL (default). It will use the annotation information from input.}

\item{intensity}{'PeakArea'(default) uses not normalized peak area.
'NormalizedPeakArea' uses peak area normalized by Spectronaut}

\item{filter_with_Qvalue}{TRUE(default) will filter out the intensities that
have greater than qvalue_cutoff in EG.Qvalue column. Those intensities will
be replaced with zero and will be considered as censored missing values for
imputation purpose.}

\item{qvalue_cutoff}{Cutoff for EG.Qvalue. default is 0.01.}

\item{useUniquePeptide}{TRUE(default) removes peptides that are assigned for
more than one proteins. We assume to use unique peptide for each protein.}

\item{fewMeasurements}{'remove' (default) will remove the features that have
1 or 2 measurements across runs.}

\item{removeProtein_with1Feature}{TRUE will remove the proteins which have
only 1 feature, which is the combination of peptide, precursor charge,
fragment and charge. FALSE is default.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple
measurements for certain feature and certain run, use highest or sum of
multiple intensities.}

\item{which.Conditions}{list of conditions to format into MSstatsPTM format.
If "all" all conditions will be used. Default is "all".}
}
\description{
Takes as as input both raw LiP and Trp outputs from Spectronautt.
}
\examples{
# Output datasets of Spectronaut
head(LiPRawData)
head(TrPRawData)

fasta_path <- "../inst/extdata/ExampleFastaFile.fasta"

MSstatsLiP_data <- SpectronauttoMSstatsLiPFormat(LiPRawData,
                                                 fasta_path,
                                                 TrPRawData)

}
