% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trypticHistogramLiP.R
\name{trypticHistogramLiP}
\alias{trypticHistogramLiP}
\title{Histogram of Half vs Fully tryptic peptides. Calculates proteotypicity,
and then uses calcualtions in histogram.}
\usage{
trypticHistogramLiP(data, fasta)
}
\description{
Histogram of Half vs Fully tryptic peptides. Calculates proteotypicity,
and then uses calcualtions in histogram.
}
