% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trypticHistogramLiP.R
\name{trypticHistogramLiP}
\alias{trypticHistogramLiP}
\title{Histogram of Half vs Fully tryptic peptides. Calculates proteotypicity,
and then uses calcualtions in histogram.}
\usage{
trypticHistogramLiP(
  data,
  fasta,
  x.axis.size = 10,
  y.axis.size = 10,
  legend.size = 10
)
}
\description{
Histogram of Half vs Fully tryptic peptides. Calculates proteotypicity,
and then uses calcualtions in histogram.
}
